prevent modify

+++ app/Http/Controllers/AccountGroupController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 use App\Models\AccountGroup;
+use App\Models\AccountHead;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Database\QueryException;
 use Illuminate\Http\JsonResponse;
@@ -57,6 +58,11 @@ class AccountGroupController extends Controller
                 ], 422);
             }
             $validated = $validator->validated();
+
+            if ($this->checkIfUsed($id))
+                return response()->json(['error' => 'Cannot not modify. The item has already been used'], 406);
+
+
             $group->update($validated);
             return response()->json($group);
         } catch (ModelNotFoundException $e) {
@@ -101,16 +107,7 @@ class AccountGroupController extends Controller
                     'errors' => $validator->errors()
                 ], 422);
             }
-
             $validated = $validator->validated();
-            if (!empty($validated['is_primary'])) {
-                AccountGroup::where('company_id', $validated['company_id'])
-                    ->where('is_primary', true)
-                    ->update(['is_primary' => false]);
-            }
-
-            $validated['is_primary'] = $validated['is_primary'] ?? false;
-
             $group = AccountGroup::create($validated);
             return response()->json($group, 201);
         } catch (ModelNotFoundException $e) {
@@ -142,6 +139,9 @@ class AccountGroupController extends Controller
     public function destroy($id): JsonResponse
     {
         try {
+            if ($this->checkIfUsed($id))
+                return response()->json(['error' => 'Cannot not modify. The item has already been used'], 406);
+
             $group = AccountGroup::findOrFail($id);
             $group->delete();
             return response()->json(['message' => 'Account Group deleted!!']);
@@ -152,5 +152,14 @@ class AccountGroupController extends Controller
             \Log::error($e);
             return response()->json(['error' => 'An unexpected error occurred!!'], 500);
         }
+    }
+
+    private function checkIfUsed($id): bool
+    {
+        if (AccountHead::where('account_group_id', $id)->first()) {
+            return true;
+        }
+        return false;
+
     }
 }
